Reworked replacenator function to remove blank spans, fixes #2
The replacenator function used blank spans to add HTML to text nodes.
Unfortunately, it didn't remove the spans after they had served their
purpose. (To be fair, it turned out to be non-obvious how to do this.) I
changed the replacenator function to use jQuery to create the span, then
an easy jQuery pattern to remove the span once it was done.

$span.contents().unwrap();

Works a charm.