Added robust debugging output and added match sort
Debugging
=========

Working with the team that provides the API has shown that there are
often issues with figuring out exactly why the plugin is responding to
sample text in various ways. A famous name isn't being matched: is the
API not matching it? Is the plugin rejecting the match? With two
separate systems working and two separate teams in charge of the
systems, it makes troubleshooting any identified issues challenging and
time-consuming.

I've added a toggle that will allow any developer to toggle on robust
output to the console of the information Interlinks receives and the
decision it makes regarding those matches.

Just set DEBUG_INTERLINKS = true in the console (a global variable), and
then use the Interlinks button.

If I can refine the debugging output to be useful in any way, let me
know.

Match Sorting by Match Offset
=============================

Sometimes the first instance of an entity was being skipped and a
later instance was matched. Why? It turns out that the matches returned
by the API weren't sorted by their location in the submitted text. I've
added a sort function on the plugin side that sorts returned matches by
their offset position so that if any match is rejected because it's
already present, the original should be the first one in the source
text.